dashboard: fix subprocesses blocking the event loop
- break apart the util module
- adds a new util to run subprocesses with asyncio
